Demonstrate RepeatedQuery Listeners
A fully reproducible example of our desired functionality using
`ExecuteListener`. The same was attempted using `DiagnosticListener`,
but with no success. The consuming code needs access to
`DiagnosticsDataSource` to proxy the connection, but it is
package-private at this time.

The idea is to hook the listener into every test that interacts with the
db via JUnit 5's Extension. That is omitted from this example for
simplicity's sake.